Move deleted files to Hadoop trash if configured (#14501)
As of now, the HadoopFileIO uses the Java delete
API, which always skips using a configured trash
directory. If the table's hadoop configuration
has trash enabled, we should use it.

We should also only do this for implementations
where trashing files is acceptable. In our case,
this is the LocalFileSystem and the
DistributedFileSystem.

Co-authored-by: Jordan Epstein <jordan.epstein@imc.com>